--- a/tests/build/test_utils.py
+++ b/tests/build/test_utils.py
@@ -46,6 +46,8 @@ def test_process_launch_training_args(job_config):
         training_args,
         tune_config,
         merge_model,
+        _,
+        _,
     ) = process_launch_training_args(job_config_copy)
     assert str(model_args.torch_dtype) == "torch.bfloat16"
     assert data_args.dataset_text_field == "output"
@@ -59,7 +61,7 @@ def test_process_launch_training_args_defaults(job_config):
     assert "torch_dtype" not in job_config_defaults
     assert job_config_defaults["use_flash_attn"] is False
     assert "save_strategy" not in job_config_defaults
-    model_args, _, training_args, _, _ = process_launch_training_args(
+    model_args, _, training_args, _, _, _, _ = process_launch_training_args(
         job_config_defaults
     )
     assert str(model_args.torch_dtype) == "torch.bfloat16"
@@ -70,13 +72,17 @@ def test_process_launch_training_args_defaults(job_config):
 def test_process_launch_training_args_peft_method(job_config):
     job_config_pt = copy.deepcopy(job_config)
     job_config_pt["peft_method"] = "pt"
-    _, _, _, tune_config, merge_model = process_launch_training_args(job_config_pt)
+    _, _, _, tune_config, merge_model, _, _ = process_launch_training_args(
+        job_config_pt
+    )
     assert isinstance(tune_config, PromptTuningConfig)
     assert merge_model is False
 
     job_config_lora = copy.deepcopy(job_config)
     job_config_lora["peft_method"] = "lora"
-    _, _, _, tune_config, merge_model = process_launch_training_args(job_config_lora)
+    _, _, _, tune_config, merge_model, _, _ = process_launch_training_args(
+        job_config_lora
+    )
     assert isinstance(tune_config, LoraConfig)
     assert merge_model is True